nexus-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.6.3
reportlab>=4.1.0
weasyprint>=61.2
python-multipart>=0.0.9
httpx>=0.27.0
redis>=5.0.1
rq>=1.16.1
PyJWT>=2.8.0
python-magic>=0.4.27
lxml>=5.1.0
boto3>=1.34.69
botocore>=1.34.69
html2text>=2024.2.26
email-validator>=2.0.0
openai>=1.0.0
Pillow>=10.0.0
matplotlib
pandas
numpy
